ai_model/requirements.txt
numpy
torch>=1.4.0
torchvision>=0.5.0
pyyaml
pandas
faiss-cpu
opencv-python
ultralytics
Pillow
scikit-learn
matplotlib
tqdm
flask
flask-cors
streamlit
koreanize-matplotlib

